chore(refactor): router -> endpoint
chore(refactor): tf vars
chore(docs): Update readme
